Move graph store into worker directory
