+++ arangod/Pregel/Algos/HITS.cpp
@@ -25,7 +25,6 @@
 #include <cmath>
 #include "Pregel/Aggregator.h"
 #include "Pregel/Algorithm.h"
-#include "Pregel/GraphStore.h"
 #include "Pregel/IncomingCache.h"
 #include "Pregel/MasterContext.h"
 #include "Pregel/VertexComputation.h"
