+++ arangod/Pregel/Algos/SSSP.cpp
@@ -23,7 +23,6 @@
 
 #include "SSSP.h"
 #include "Pregel/Algorithm.h"
-#include "Pregel/GraphStore.h"
 #include "Pregel/IncomingCache.h"
 #include "Pregel/VertexComputation.h"
