+++ arangod/Pregel/OutgoingCache.h
@@ -27,7 +27,6 @@
 #include "Cluster/ClusterInfo.h"
 #include "VocBase/voc-types.h"
 
-#include "Pregel/GraphStore.h"
 #include "Pregel/MessageCombiner.h"
 #include "Pregel/MessageFormat.h"
 #include "Pregel/Worker/WorkerConfig.h"
