+++ arangod/Pregel/Worker/CMakeLists.txt
@@ -1,3 +1,4 @@
 target_sources(arango_pregel PRIVATE
 	Worker.cpp
-	WorkerConfig.cpp)
+	WorkerConfig.cpp
+	GraphStore.cpp)
