+++ arangod/Pregel/Worker/GraphStore.cpp
